Move macros and declarations into a header file
